python/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
dnspython==1.16.0
ez-setup==0.9
feedparser==5.2.1
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
moviepy==1.0.2
music21==5.7.2
mysql-connector==2.2.9
mysql-connector-python==8.0.19
numpy==1.18.2
Pillow==7.1.1
proglog==0.1.9
protobuf==3.6.1
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
slackweb==1.0.5
tqdm==4.45.0
unroll==0.1.0
urllib3==1.25.9
